moved the mscolab-coupled helpers to mscolab.auth
