+++ mslib/mscolab/blueprints/auth/__init__.py
@@ -37,11 +37,12 @@ from flask.wrappers import Response
 from saml2 import BINDING_HTTP_REDIRECT, BINDING_HTTP_POST
 from saml2.metadata import create_metadata_string
 
+from mslib.mscolab.auth import check_login, register_user, generate_confirmation_token, create_or_update_idp_user, \
+    get_idp_entity_id, confirm_token
 from mslib.mscolab.conf import setup_saml2_backend
 from mslib.mscolab.forms import ResetPasswordForm, ResetRequestForm
 from mslib.mscolab.models import User
-from mslib.utils.auth import check_login, register_user, generate_confirmation_token, send_email, confirm_token, \
-    get_idp_entity_id, create_or_update_idp_user
+from mslib.utils.auth import send_email
 
 AUTH_BP = Blueprint('auth', __name__, template_folder='templates')
