+++ tests/_test_mscolab/test_server_auth_required.py
@@ -32,7 +32,7 @@ mscolab_settings.enable_basic_http_authentication = True
 try:
     from mslib.mscolab.server import authfunc, verify_pw, initialize_managers
     from mslib.mscolab.blueprints.auth import get_auth_token
-    from mslib.utils.auth import register_user
+    from mslib.mscolab.auth import register_user
 except ImportError:
     pytest.skip("this test runs only by an explicit call "
                 "e.g. pytest tests/_test_mscolab/test_server_auth_required.py", allow_module_level=True)
